rewrite the functionality to save flashcard sets to firebase as a hook
